CO-0022: Add profile view model and refactor profile view
