--- DubDubGrub/Views/Screens/ProfileView.swift
@@ -1,229 +1,0 @@
-//
-//  ProfileView.swift
-//  DubDubGrub
-//
-//  Created by Oscar Cristaldo on 2022-08-09.
-//
-
-import SwiftUI
-import CloudKit
-
-struct ProfileView: View {
-    
-    @State private var firstName            = ""
-    @State private var lastName             = ""
-    @State private var companyName          = ""
-    @State private var bio                  = ""
-    @State private var avatar               = PlaceholderImage.avatar
-    @State private var isShowingPhotoPicker = false
-    @State private var alertItem: AlertItem?
-    @FocusState var dismissKeyboard: Bool
-    
-    var body: some View {
-        VStack {
-            ZStack {
-                NameBackgroundView()
-                
-                HStack(spacing: 16) {
-                    ZStack {
-                        AvatarView(image: avatar, size: 84)
-                        EditImageView()
-                    }
-                    .padding(.leading, 12)
-                    .onTapGesture {
-                        isShowingPhotoPicker = true
-                    }
-                    
-                    VStack(spacing: 1) {
-                        TextField("First Name", text: $firstName).profileNameStyle()
-                        TextField("Last Name", text: $lastName).profileNameStyle()
-                        TextField("Company Name", text: $companyName)
-                    }
-                    .padding(.trailing, 16)
-                }
-                .padding()
-            }
-            
-            VStack(alignment: .leading, spacing: 8) {
-                CharacterRemainView(currentCount: bio.count)
-                
-                TextEditor(text: $bio)
-                    .focused($dismissKeyboard)
-                    .toolbar {
-                        ToolbarItem(placement: .keyboard) {
-                            HStack {
-                                Spacer()
-                                Button {
-                                    dismissKeyboard.toggle()
-                                } label: {
-                                    Image(systemName: "keyboard.chevron.compact.down")
-                                        .foregroundColor(Color("AccentColor"))
-                                }
-                            }
-                        }
-                    }
-                    .frame(height: 100)
-                    .overlay(RoundedRectangle(cornerRadius: 8)
-                        .stroke(Color.secondary, lineWidth: 1))
-                
-            }
-            .padding(.horizontal, 20)
-            
-            Spacer()
-            
-            Button {
-                //                                createProfile()
-            } label: {
-                DDGButton(title: "Create Profile")
-            }
-            .padding(.bottom)
-        }
-        .navigationTitle("Profile")
-        .onAppear{ getProfile() }
-        .alert(item: $alertItem, content: { alertItem in
-            Alert(title: alertItem.title, message: alertItem.message, dismissButton: alertItem.dismissButton)
-        })
-        .sheet(isPresented: $isShowingPhotoPicker) {
-            PhotoPicker(image: $avatar)
-        }
-    }
-    
-    func isValidProfile() -> Bool {
-        guard !firstName.isEmpty,
-              !lastName.isEmpty,
-              !companyName.isEmpty,
-              !bio.isEmpty,
-              avatar != PlaceholderImage.avatar,
-              bio.count < 100 else { return false }
-        return true
-    }
-    
-    func createProfile() {
-        guard isValidProfile() else {
-            alertItem = AlertContext.invalidProfile
-            return
-        }
-        
-        let profileRecord = CKRecord(recordType: RecordType.profile)
-        profileRecord[DDGProfile.kFirstName]    = firstName
-        profileRecord[DDGProfile.kLastName]     = lastName
-        profileRecord[DDGProfile.kCompanyName]  = companyName
-        profileRecord[DDGProfile.kBio]          = bio
-        profileRecord[DDGProfile.kAvatar]       = avatar.convertToCKAsset()
-        
-        CKContainer.default().fetchUserRecordID { recordID, error in
-            guard let recordID = recordID, error == nil else {
-                print(error!.localizedDescription)
-                return
-            }
-            
-            CKContainer.default().privateCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
-                guard let userRecord = userRecord, error == nil else {
-                    print(error!.localizedDescription)
-                    return
-                }
-                
-                userRecord["userProfile"] = CKRecord.Reference(recordID: profileRecord.recordID, action: .none)
-                
-                let operation = CKModifyRecordsOperation(recordsToSave: [userRecord, profileRecord])
-                operation.qualityOfService = .userInteractive
-                
-                operation.modifyRecordsResultBlock = { result in
-                    switch result {
-                    case .success:
-                        print("Succesfully created and uploaded profile to CloudKit")
-                    case .failure(let error):
-                        print(error.localizedDescription)
-                    }
-                }
-                
-                CKContainer.default().publicCloudDatabase.add(operation)
-            }
-            
-        }
-    }
-    
-    func getProfile() {
-        CKContainer.default().fetchUserRecordID { recordID, error in
-            guard let recordID = recordID, error == nil else {
-                print(error!.localizedDescription)
-                return
-            }
-            
-            CKContainer.default().privateCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
-                guard let userRecord = userRecord, error == nil else {
-                    print(error!.localizedDescription)
-                    return
-                }
-                
-                let profileReference = userRecord["userProfile"] as! CKRecord.Reference
-                let profileRecordID = profileReference.recordID
-                
-                CKContainer.default().publicCloudDatabase.fetch(withRecordID: profileRecordID) { profileRecord, error in
-                    guard let profileRecord = profileRecord, error == nil else {
-                        print(error!.localizedDescription)
-                        return
-                    }
-                    
-                    DispatchQueue.main.async {
-                        let profile = DDGProfile(record: profileRecord)
-                        firstName   = profile.firstName
-                        lastName    = profile.lastName
-                        companyName = profile.companyName
-                        bio         = profile.bio
-                        avatar      = profile.createAvatarImage()
-                    }
-                }
-            }
-        }
-    }
-}
-
-struct ProfileView_Previews: PreviewProvider {
-    static var previews: some View {
-        NavigationView {
-            ProfileView()
-        }
-        
-    }
-}
-
-struct NameBackgroundView: View {
-    var body: some View {
-        Color(uiColor: .secondarySystemBackground)
-            .frame(height: 130)
-            .cornerRadius(12)
-            .padding(.horizontal)
-    }
-}
-
-struct EditImageView: View {
-    var body: some View {
-        Image(systemName: "square.and.pencil")
-            .resizable()
-            .scaledToFit()
-            .frame(width: 14, height: 14)
-            .foregroundColor(.white)
-            .offset(y: 30)
-    }
-}
-
-struct CharacterRemainView: View {
-    
-    var currentCount: Int
-    
-    var body: some View {
-        Text("Bio: ")
-            .font(.callout)
-            .foregroundColor(.secondary)
-        +
-        Text("\(100 - currentCount)")
-            .bold()
-            .font(.callout)
-            .foregroundColor(currentCount <= 100 ? Color("AccentColor") : Color(uiColor: .systemPink))
-        +
-        Text(" Characters Remain")
-            .font(.callout)
-            .foregroundColor(.secondary)
-    }
-}
