+++ DubDubGrub/Views/Screens/ProfileView/ProfileView.swift
@@ -1,0 +1,137 @@
+//
+//  ProfileView.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-08-09.
+//
+
+import SwiftUI
+import CloudKit
+
+struct ProfileView: View {
+    
+    @StateObject private var viewModel = ProfileViewModel()
+    @FocusState var dismissKeyboard: Bool
+    
+    var body: some View {
+        VStack {
+            ZStack {
+                NameBackgroundView()
+                
+                HStack(spacing: 16) {
+                    ZStack {
+                        AvatarView(image: viewModel.avatar, size: 84)
+                        EditImageView()
+                    }
+                    .padding(.leading, 12)
+                    .onTapGesture { viewModel.isShowingPhotoPicker = true }
+                    
+                    VStack(spacing: 1) {
+                        TextField("First Name", text: $viewModel.firstName).profileNameStyle()
+                        TextField("Last Name", text: $viewModel.lastName).profileNameStyle()
+                        TextField("Company Name", text: $viewModel.companyName)
+                    }
+                    .padding(.trailing, 16)
+                }
+                .padding()
+            }
+            
+            VStack(alignment: .leading, spacing: 8) {
+                CharacterRemainView(currentCount: viewModel.bio.count)
+                
+                TextEditor(text: $viewModel.bio)
+                    .focused($dismissKeyboard)
+                    .toolbar {
+                        ToolbarItem(placement: .keyboard) {
+                            HStack {
+                                Spacer()
+                                Button {
+                                    dismissKeyboard.toggle()
+                                } label: {
+                                    Image(systemName: "keyboard.chevron.compact.down")
+                                        .foregroundColor(Color("AccentColor"))
+                                }
+                            }
+                        }
+                    }
+                    .frame(height: 100)
+                    .overlay(RoundedRectangle(cornerRadius: 8)
+                        .stroke(Color.secondary, lineWidth: 1))
+                
+            }
+            .padding(.horizontal, 20)
+            
+            Spacer()
+            
+            Button {
+                //                                createProfile()
+            } label: {
+                DDGButton(title: "Create Profile")
+            }
+            .padding(.bottom)
+        }
+        .navigationTitle("Profile")
+        .onAppear{ viewModel.getProfile() }
+        .alert(item: $viewModel.alertItem, content: { alertItem in
+            Alert(title: alertItem.title, message: alertItem.message, dismissButton: alertItem.dismissButton)
+        })
+        .sheet(isPresented: $viewModel.isShowingPhotoPicker) {
+            PhotoPicker(image: $viewModel.avatar)
+        }
+    }
+    
+    
+    
+    
+    
+    
+}
+
+struct ProfileView_Previews: PreviewProvider {
+    static var previews: some View {
+        NavigationView {
+            ProfileView()
+        }
+        
+    }
+}
+
+struct NameBackgroundView: View {
+    var body: some View {
+        Color(uiColor: .secondarySystemBackground)
+            .frame(height: 130)
+            .cornerRadius(12)
+            .padding(.horizontal)
+    }
+}
+
+struct EditImageView: View {
+    var body: some View {
+        Image(systemName: "square.and.pencil")
+            .resizable()
+            .scaledToFit()
+            .frame(width: 14, height: 14)
+            .foregroundColor(.white)
+            .offset(y: 30)
+    }
+}
+
+struct CharacterRemainView: View {
+    
+    var currentCount: Int
+    
+    var body: some View {
+        Text("Bio: ")
+            .font(.callout)
+            .foregroundColor(.secondary)
+        +
+        Text("\(100 - currentCount)")
+            .bold()
+            .font(.callout)
+            .foregroundColor(currentCount <= 100 ? Color("AccentColor") : Color(uiColor: .systemPink))
+        +
+        Text(" Characters Remain")
+            .font(.callout)
+            .foregroundColor(.secondary)
+    }
+}
