+++ DubDubGrub/Views/Screens/ProfileView/ProfileViewModel.swift
@@ -1,0 +1,115 @@
+//
+//  ProfileViewModel.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-09-11.
+//
+
+import CloudKit
+
+final class ProfileViewModel: ObservableObject {
+    
+    @Published var firstName            = ""
+    @Published var lastName             = ""
+    @Published var companyName          = ""
+    @Published var bio                  = ""
+    @Published var avatar               = PlaceholderImage.avatar
+    @Published var isShowingPhotoPicker = false
+    @Published var alertItem: AlertItem?
+    
+    func isValidProfile() -> Bool {
+        guard !firstName.isEmpty,
+              !lastName.isEmpty,
+              !companyName.isEmpty,
+              !bio.isEmpty,
+              avatar != PlaceholderImage.avatar,
+              bio.count < 100 else { return false }
+        return true
+    }
+    
+    func getProfile() {
+        CKContainer.default().fetchUserRecordID { recordID, error in
+            guard let recordID = recordID, error == nil else {
+                print(error!.localizedDescription)
+                return
+            }
+            
+            CKContainer.default().privateCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
+                guard let userRecord = userRecord, error == nil else {
+                    print(error!.localizedDescription)
+                    return
+                }
+                
+                let profileReference = userRecord["userProfile"] as! CKRecord.Reference
+                let profileRecordID = profileReference.recordID
+                
+                CKContainer.default().publicCloudDatabase.fetch(withRecordID: profileRecordID) { profileRecord, error in
+                    guard let profileRecord = profileRecord, error == nil else {
+                        print(error!.localizedDescription)
+                        return
+                    }
+                    
+                    DispatchQueue.main.async { [self] in
+                        let profile = DDGProfile(record: profileRecord)
+                        firstName   = profile.firstName
+                        lastName    = profile.lastName
+                        companyName = profile.companyName
+                        bio         = profile.bio
+                        avatar      = profile.createAvatarImage()
+                    }
+                }
+            }
+        }
+    }
+    
+    func createProfile() {
+        guard isValidProfile() else {
+            alertItem = AlertContext.invalidProfile
+            return
+        }
+        
+        let profileRecord = createProfileRecord() 
+        
+        CKContainer.default().fetchUserRecordID { recordID, error in
+            guard let recordID = recordID, error == nil else {
+                print(error!.localizedDescription)
+                return
+            }
+            
+            CKContainer.default().privateCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
+                guard let userRecord = userRecord, error == nil else {
+                    print(error!.localizedDescription)
+                    return
+                }
+                
+                userRecord["userProfile"] = CKRecord.Reference(recordID: profileRecord.recordID, action: .none)
+                
+                let operation = CKModifyRecordsOperation(recordsToSave: [userRecord, profileRecord])
+                operation.qualityOfService = .userInteractive
+                
+                operation.modifyRecordsResultBlock = { result in
+                    switch result {
+                    case .success:
+                        print("Succesfully created and uploaded profile to CloudKit")
+                    case .failure(let error):
+                        print(error.localizedDescription)
+                    }
+                }
+                
+                CKContainer.default().publicCloudDatabase.add(operation)
+            }
+            
+        }
+    }
+    
+    private func createProfileRecord() -> CKRecord {
+        let profileRecord = CKRecord(recordType: RecordType.profile)
+        profileRecord[DDGProfile.kFirstName]    = firstName
+        profileRecord[DDGProfile.kLastName]     = lastName
+        profileRecord[DDGProfile.kCompanyName]  = companyName
+        profileRecord[DDGProfile.kBio]          = bio
+        profileRecord[DDGProfile.kAvatar]       = avatar.convertToCKAsset()
+        
+        return profileRecord
+    }
+}
